Consolidate unit tests into core and sanitizer test modules
Reorganized scattered test files into two main test modules:
- test_core.py: Contains general functionality tests from test_autotune_add,
  test_config, test_trace_add_clients, and test_wrapper
- test_sanitizer.py: Consolidates all sanitizer-related tests including
  indirect_load, symbolic_execution, and null_sanitizer tests

This reorganization improves test maintainability by grouping related tests
together and reducing the number of test files from 12 to 2.
