--- tests/indirect_load/test_const_dtype_inference.py
@@ -1,20 +1,0 @@
-import triton.language as tl
-from triton_viz.clients.sanitizer.sanitizer import SymbolicExpr
-
-
-def test_const_dtype_inference():
-    x = SymbolicExpr.from_value((1, 2, 3))
-    y = SymbolicExpr.from_value((1, 2, 3))
-    z = SymbolicExpr("add", x, y)
-
-    assert x.dtype_tt == tl.int32
-    assert y.dtype_tt == tl.int32
-    assert z.dtype_tt == tl.int32
-
-    expected_tree = (
-        "\n"
-        "add [dtype=int32]\n"
-        "├── lhs: const=(1, 2, 3) [dtype=int32]\n"
-        "└── rhs: const=(1, 2, 3) [dtype=int32]"
-    )
-    assert z.to_tree_str() == expected_tree
